+++ packages/bluesky_cli/CHANGELOG.md
@@ -10,6 +10,7 @@
 - Supported `add-list-item` command. ([#393](https://github.com/myConsciousness/atproto.dart/issues/393))
 - Supported `muting-lists` command. ([#390](https://github.com/myConsciousness/atproto.dart/issues/390))
 - Supported `mute-list` command. ([#394](https://github.com/myConsciousness/atproto.dart/issues/394))
+- Supported `unmute-list` command. ([#395](https://github.com/myConsciousness/atproto.dart/issues/394))
 
 ## v0.1.2